+++ CHANGES.md
@@ -3,6 +3,7 @@ Change Log
 - Version 3.0.8 (pending)
     - Fixed bug where you couldn't add attendees to an at con order with a zero-cost attendee
     - Fixed bug where completing a zero-cost order wouldn't navigate to print badge screen
+    - Fixed bug where birthdays on 12/31 would show up in the wrong year on check in screen
 
 - Version 3.0.7 (7/26/2018)
     - Added Edit button on check in views